Skip teardown in GH & remove stdio: []

+++ Jest/globalTeardown.js
@@ -4,9 +4,13 @@ const execSync = require("child_process").execSync;
 const parentFolderPath = require("../Utils/folderPaths").parentFolderPath();
 
 module.exports = function () {
+  //skip when running in GitHub
+  if (process.env.GITHUB_JOB !== undefined) return;
+
+  // start teardown
   process.chdir(os.homedir());
 
-  execSync("npm uninstall prettier sort-package-json --global", { stdio: [] });
+  execSync("npm uninstall prettier sort-package-json --global");
 
   fs.rmSync(parentFolderPath, {
     recursive: true,
